+++ chat-service/.env
@@ -1,0 +1,8 @@
+DATABASE_URL=mysql+aiomysql://root:password@localhost:3306/chat_service
+REDIS_URL=redis://localhost:6379/0
+AUTH_SERVICE_URL=mysql+aiomysql://root:password@localhost:3306/auth_service
+SECRET_KEY=AVGTEAMVNLAW
+CHROMA_COLLECTION_NAME=chat-service
+CHROMA_URL=http://localhost:3307
+DB_NAME=law_service
+OPENAI_API_KEY=sk-proj-V01M85L8vYNlxT6kBmCpT3BlbkFJNsgTUeevlK0JoRzck4xq
