+++ chat-service/pom.xml
@@ -84,6 +84,16 @@
             <artifactId>netty-socketio</artifactId>
             <version>2.0.3</version>
         </dependency>
+        <dependency>
+            <groupId>org.springframework.kafka</groupId>
+            <artifactId>spring-kafka</artifactId>
+        </dependency>
+        <!-- https://mvnrepository.com/artifact/net.java.dev.jna/jna -->
+        <dependency>
+            <groupId>org.springframework.integration</groupId>
+            <artifactId>spring-integration-kafka</artifactId>
+            <version>6.3.5</version>
+        </dependency>
     </dependencies>
 
     <build>
