+++ chat-service/src/main/java/fit/hutech/service/chatservice/DTO/ChatResponse.java
@@ -1,0 +1,14 @@
+package fit.hutech.service.chatservice.DTO;
+
+import fit.hutech.service.chatservice.models.AnswerResult;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class ChatResponse {
+    private String messageId;
+    private AnswerResult answer;
+}
