+++ chat-service/src/main/java/fit/hutech/service/chatservice/services/ArticleService.java
@@ -2,13 +2,29 @@ package fit.hutech.service.chatservice.services;
 
 import fit.hutech.service.chatservice.DTO.ArticleDTO;
 import fit.hutech.service.chatservice.models.Article;
-import org.springframework.data.domain.Page;
+import fit.hutech.service.chatservice.repositories.ArticleRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
 
-public interface ArticleService {
-    public List<ArticleDTO> getArticlesWithRelatedInfo();
-    public void save(Article article);
-    public Article getArticleById(String id);
+@Service
+@RequiredArgsConstructor
+public class ArticleService {
+
+    private final ArticleRepository articleRepository ;
+
+
+    public List<ArticleDTO> getArticlesWithRelatedInfo() {
+        return articleRepository.getArticlesWithRelatedInfo();
+    }
+
+    public void save(Article article) {
+        articleRepository.save(article);
+    }
+
+
+    public Article getArticleById(String id) {
+        return articleRepository.findById(id).orElse(null);
+    }
 }
