--- chat-service/src/main/java/fit/hutech/service/chatservice/services/SocketService.java
@@ -1,9 +1,0 @@
-package fit.hutech.service.chatservice.services;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public interface SocketService {
-    public void Test();
-    public void handleSocketService();
-}
