+++ chat-service/src/main/java/fit/hutech/service/chatservice/services/VbqpplService.java
@@ -1,9 +1,21 @@
 package fit.hutech.service.chatservice.services;
 
 import fit.hutech.service.chatservice.models.Vbqppl;
+import fit.hutech.service.chatservice.repositories.VbqpplRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
 
 import java.util.List;
 
-public interface VbqpplService{
-    public List<Vbqppl> getAll();
+@Service
+@RequiredArgsConstructor
+public class VbqpplService {
+
+    private final VbqpplRepository vbqpplRepository;
+
+    public List<Vbqppl> getAll() {
+        return vbqpplRepository.findAll();
+    }
+
+
 }
