+++ chat-service/src/main/resources/application.properties
@@ -2,7 +2,7 @@ spring.application.name=chat-service
 server.port=9003
 
 ## PERSISTENCE
-spring.datasource.url=jdbc:mysql://localhost:3306/law_service?useSSL=false&allowPublicKeyRetrieval=true
+spring.datasource.url=jdbc:mysql://localhost:4000/law_service?useSSL=false&allowPublicKeyRetrieval=true
 spring.datasource.username=root
 spring.datasource.password=password
 spring.jpa.hibernate.ddl-auto=update
@@ -12,3 +12,16 @@ logging.level.org.hibernate=ERROR
 # discovery server
 eureka.client.serviceUrl.defaultZone=http://localhost:8761/eureka/
 #eureka.client.enabled=false
+
+
+# kafka
+spring.kafka.bootstrap-servers= localhost:9092
+# kafka consumer
+spring.kafka.consumer.group-id= chat-group
+spring.kafka.consumer.key-deserializer= org.apache.kafka.common.serialization.StringDeserializer
+spring.kafka.consumer.value-deserializer=org.apache.kafka.common.serialization.StringDeserializer
+
+# kafka producer
+spring.kafka.producer.key-serializer= org.apache.kafka.common.serialization.StringSerializer
+spring.kafka.producer.value-serializer=org.apache.kafka.common.serialization.StringSerializer
+
