+++ front-end/website/src/common/constants/constant.js
@@ -1,0 +1,5 @@
+const MAX_RECONNECT_ATTEMPTS = 5;
+const RECONNECT_INTERVAL = 3000;
+const MESSAGE_TIMEOUT = 30000;
+
+export { MAX_RECONNECT_ATTEMPTS, RECONNECT_INTERVAL, MESSAGE_TIMEOUT };
